moves readable, writable interfaces into single module
