+++ allensdk/brain_observatory/behavior/data_objects/cell_specimen_table.py
@@ -6,16 +6,10 @@ from pynwb.ophys import OpticalChannel, ImageSegmentation
 import allensdk.brain_observatory.roi_masks as roi
 from allensdk.brain_observatory.behavior.data_objects import DataObject
 from allensdk.brain_observatory.behavior.data_objects.base \
-    .readable_interfaces.json_readable_interface import \
-    JsonReadableInterface
-from allensdk.brain_observatory.behavior.data_objects.base \
-    .readable_interfaces.lims_readable_interface import \
-    LimsReadableInterface
-from allensdk.brain_observatory.behavior.data_objects.base \
-    .readable_interfaces.nwb_readable_interface import \
-    NwbReadableInterface
-from allensdk.brain_observatory.behavior.data_objects.base \
-    .writable_interfaces.nwb_writable_interface import \
+    .readable_interfaces import \
+    JsonReadableInterface, LimsReadableInterface, NwbReadableInterface
+from allensdk.brain_observatory.behavior.data_objects.base\
+    .writable_interfaces import \
     NwbWritableInterface
 from allensdk.brain_observatory.behavior.data_objects.metadata\
     .behavior_ophys_metadata import \
