+++ allensdk/brain_observatory/behavior/data_objects/metadata/behavior_metadata/behavior_session_uuid.py
@@ -5,12 +5,9 @@ from pynwb import NWBFile
 
 from allensdk.brain_observatory.behavior.data_files import StimulusFile
 from allensdk.brain_observatory.behavior.data_objects import DataObject
-from allensdk.brain_observatory.behavior.data_objects.base\
-    .readable_interfaces.nwb_readable_interface import \
-    NwbReadableInterface
-from allensdk.brain_observatory.behavior.data_objects.base.readable_interfaces\
-    .stimulus_file_readable_interface import \
-    StimulusFileReadableInterface
+from allensdk.brain_observatory.behavior.data_objects.base \
+    .readable_interfaces import \
+    NwbReadableInterface, StimulusFileReadableInterface
 
 
 class BehaviorSessionUUID(DataObject, StimulusFileReadableInterface,
