+++ allensdk/brain_observatory/behavior/data_objects/metadata/ophys_experiment_metadata/mesoscope_experiment_metadata/imaging_plane_group.py
@@ -2,14 +2,8 @@ from pynwb import NWBFile
 
 from allensdk.brain_observatory.behavior.data_objects import DataObject
 from allensdk.brain_observatory.behavior.data_objects.base \
-    .readable_interfaces.json_readable_interface import \
-    JsonReadableInterface
-from allensdk.brain_observatory.behavior.data_objects.base \
-    .readable_interfaces.lims_readable_interface import \
-    LimsReadableInterface
-from allensdk.brain_observatory.behavior.data_objects.base \
-    .readable_interfaces.nwb_readable_interface import \
-    NwbReadableInterface
+    .readable_interfaces import \
+    JsonReadableInterface, LimsReadableInterface, NwbReadableInterface
 from allensdk.internal.api import PostgresQueryMixin
