+++ allensdk/brain_observatory/behavior/data_objects/metadata/ophys_experiment_metadata/ophys_experiment_metadata.py
@@ -4,15 +4,9 @@ from typing import Optional
 from pynwb import NWBFile
 
 from allensdk.brain_observatory.behavior.data_objects import DataObject
-from allensdk.brain_observatory.behavior.data_objects.base\
-    .readable_interfaces.internal_readable_interface import \
-    InternalReadableInterface
-from allensdk.brain_observatory.behavior.data_objects.base\
-    .readable_interfaces.json_readable_interface import \
-    JsonReadableInterface
-from allensdk.brain_observatory.behavior.data_objects.base\
-    .readable_interfaces.nwb_readable_interface import \
-    NwbReadableInterface
+from allensdk.brain_observatory.behavior.data_objects.base \
+    .readable_interfaces import \
+    InternalReadableInterface, JsonReadableInterface, NwbReadableInterface
 from allensdk.brain_observatory.behavior.data_objects.metadata\
     .ophys_experiment_metadata.emission_lambda import \
     EmissionLambda
