+++ allensdk/brain_observatory/behavior/data_objects/stimulus_timestamps/stimulus_timestamps.py
@@ -8,32 +8,17 @@ import numpy as np
 from pynwb import NWBFile, ProcessingModule
 from pynwb.base import TimeSeries
 
-from allensdk.brain_observatory.behavior.data_objects.base\
-    .readable_interfaces.json_readable_interface import \
-    JsonReadableInterface
-from allensdk.brain_observatory.behavior.data_objects.base\
-    .readable_interfaces.lims_readable_interface import \
-    LimsReadableInterface
-from allensdk.brain_observatory.behavior.data_objects.base\
-    .readable_interfaces.nwb_readable_interface import \
-    NwbReadableInterface
-from allensdk.brain_observatory.behavior.data_objects.base.readable_interfaces\
-    .stimulus_file_readable_interface \
-    import \
-    StimulusFileReadableInterface
-from allensdk.brain_observatory.behavior.data_objects.base.readable_interfaces\
-    .sync_file_readable_interface import \
-    SyncFileReadableInterface
+from allensdk.brain_observatory.behavior.data_objects.base \
+    .readable_interfaces import \
+    JsonReadableInterface, LimsReadableInterface, NwbReadableInterface, \
+    StimulusFileReadableInterface, SyncFileReadableInterface
 from allensdk.brain_observatory.behavior.data_objects import DataObject
 from allensdk.brain_observatory.behavior.data_files import (
     StimulusFile, SyncFile
 )
-from allensdk.brain_observatory.behavior.data_objects.base\
-    .writable_interfaces.json_writable_interface import \
-    JsonWritableInterface
-from allensdk.brain_observatory.behavior.data_objects.base\
-    .writable_interfaces.nwb_writable_interface import \
-    NwbWritableInterface
+from allensdk.brain_observatory.behavior.data_objects.base \
+    .writable_interfaces import \
+    JsonWritableInterface, NwbWritableInterface
 from allensdk.brain_observatory.behavior.data_objects.stimulus_timestamps.timestamps_processing import (  # noqa: E501
     get_behavior_stimulus_timestamps, get_ophys_stimulus_timestamps
 )
